requirements.txt
fastapi
uvicorn
streamlit
scikit-learn
pandas
tensorflow
python-multipart
google-cloud-storage==2.14.0
transformers
torch
